Update all ids from Long to Int

+++ feature-pokedex/src/main/java/com/victor/feature_pokedex/data/PokedexGateway.kt
@@ -20,20 +20,20 @@ internal interface PokedexGateway {
     ): PagedResponse<NameAndUrlResponse>
 
     @GET("pokemon/{id}")
-    suspend fun getPokemon(@Path("id") pokemonId: Long): PokemonResponse
+    suspend fun getPokemon(@Path("id") pokemonId: Int): PokemonResponse
 
     @GET("type")
     suspend fun getTypeList(): PagedResponse<NameAndUrlResponse>
 
     @GET("type/{id}")
-    suspend fun getType(@Path("id") typeId: Long): TypeResponse
+    suspend fun getType(@Path("id") typeId: Int): TypeResponse
 
     @GET("generation/{generation}")
     suspend fun getGeneration(@Path("generation") generationId: Int): GenerationResponse
 
     @GET("pokemon-species/{id}")
-    suspend fun getSpecie(@Path("id") pokemonId: Long): SpecieResponse
+    suspend fun getSpecie(@Path("id") pokemonId: Int): SpecieResponse
 
     @GET("evolution-chain/{id}")
-    suspend fun getEvolutionChain(@Path("id") chainId: Long): EvolutionResponse
+    suspend fun getEvolutionChain(@Path("id") chainId: Int): EvolutionResponse
 }
